Ajout Swap

+++ Src/Metaheuristique/fonction_rand.jl
@@ -123,5 +123,5 @@ function violation_same_color(instance::Array{Array{Int32,1},1},prio::Array{Arra
 end
 
 
-generic(1000)
-generic(1000)
+generic(Int32(1000))
+generic(Int32(1000))
